+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog for ghc-lib-parser-ex
 
+## Unreleased
+- Update types in `GHC.Types.Name.Reader` for ghc-9.2.1, `Located` becomes `LocatedN`
+
 ## 0.20210501 released 2021-05-01
 - Update to `ghc-lib-0.20210501`